refactor: move matrices module into a separate package.
