+++ libs/fakeblocks/pyproject.toml
@@ -1,0 +1,25 @@
+[project]
+name = "fakeblocks"
+version = "1.0"
+description = ""
+authors = [{name = "Matthieu Ancellin", email = "matthieu.ancellin@mews-labs.com"}]
+requires-python = ">=3.8"
+dependencies = [
+   "numpy",
+   "scipy",
+   "capytaine>=3.0",
+]
+urls = {repository = "https://github.com/capytaine/capytaine"}
+readme = "README.md"
+license = {file = "LICENSE"}
+classifiers = [
+  "Programming Language :: Python",
+  "Programming Language :: Fortran",
+  "Intended Audience :: Science/Research",
+  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
+  "Topic :: Scientific/Engineering",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
