+++ libs/fakeblocks/src/fakeblocks/matrices/__init__.py
@@ -2,15 +2,15 @@
 These matrices can be nested to recursively define Hierarchical matrices.
 """
 # Copyright (C) 2017-2019 Matthieu Ancellin
-# See LICENSE file at <https://github.com/mancellin/capytaine>
+# See LICENSE file at <https://github.com/capytaine/fakeblocks>
 
-from capytaine.matrices.block import BlockMatrix
-from capytaine.matrices.block_toeplitz import (
+from fakeblocks.matrices.block import BlockMatrix
+from fakeblocks.matrices.block_toeplitz import (
     BlockToeplitzMatrix, BlockSymmetricToeplitzMatrix,
     BlockCirculantMatrix, EvenBlockSymmetricCirculantMatrix, OddBlockSymmetricCirculantMatrix,
 )
-from capytaine.matrices.builders import (
+from fakeblocks.matrices.builders import (
     cut_matrix, random_block_matrix,
     full_like, zeros_like, ones_like, identity_like,
 )
-from capytaine.matrices.low_rank import LowRankMatrix
+from fakeblocks.matrices.low_rank import LowRankMatrix
