+++ libs/fakeblocks/src/fakeblocks/matrices/low_rank.py
